+++ web-wallet/src/routes/(welcome)/setup/PasswordSetup.svelte
@@ -11,7 +11,7 @@
   export let isValid = false;
 
   /** @type {boolean} */
-  export let isToggled = false;
+  export let isToggled = true;
 
   /** @type {string} */
   let confirmPassword = "";
@@ -59,13 +59,13 @@
 </ToggleableCard>
 
 <Banner
-  title="Setting a password for your web wallet is optional."
+  title="Setting a password for your web wallet is recommended."
   variant="info"
 >
   <p>
-    Doing so allows you the convenience of accessing your wallet using a
-    password, but it weakens the overall security. Not using a password requires
-    you to input the full mnemonic to open your wallet.
+    Without a password, you will need to input your full mnemonic each time to
+    access your wallet. This could expose your mnemonic to potential threats.
+    Setting a password is highly recommended for better wallet security.
   </p>
 </Banner>