+++ web-wallet/src/routes/(welcome)/setup/restore/+page.svelte
@@ -37,7 +37,7 @@
   let isSyncCompleted = false;
 
   /** @type {boolean} */
-  let showPasswordSetup = false;
+  let showPasswordSetup = true;
 
   /** @type {boolean} */
   let isValidMnemonic = false;